pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
